Updated Rig
Fixed some skinning issues around the elbow/wrist area. It should be a bit better now
Nice catch on the stretchy issue. Since we added twist bones, I needed to apply the stretch on those as well so it should be working better now.
I added the on/off switch for the sheer geo but, like we said above with Jason, maybe we can edit it to more of a material edit.
BONUS: I added controls and on/off switch for the necklace that Jason modeled. Super simple but now you can move it.